requirements.txt
click==8.1.7
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn==0.29.0
langchain-core==0.1.52
pypdf==4.2.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2

# Optional (bonus)
celery[redis]==5.3.1
redis==5.0.10
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
